AutoGen/Projects/AI Interview Agentic System/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
pyautogen==0.2.19
openai==1.12.0
faiss-cpu==1.7.4
pypdf==3.17.4
python-docx==1.1.0
numpy==1.26.3
